requirements.txt
cryptography
cv2
matplotlib
notebook
numpy
paho-mqtt
pillow
scikit-learn
seaborn
watchdog
xgboost